feat(editor): more demo preview, consume editor package, cleanup
