+++ packages/editor/demo/lit/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Editor Web Component</title>
+    <title>lit element</title>
     <script src="./index.js" type="module"></script>
   </head>
   <body>
