+++ packages/editor/demo/react-preview/index.html
@@ -3,10 +3,10 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>editor react | standalone</title>
+    <title>editor react | with preview</title>
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/___dev-demo/standalone.tsx"></script>
+    <script type="module" src="/src/___dev-demo/main.tsx"></script>
   </body>
 </html>
