+++ packages/editor/index.html
@@ -7,11 +7,16 @@
   </head>
   <body>
     <h1>Hi there 👋</h1>
+
+    <p>Here are some live updating demos:</p>
     <ul>
+      <li><a href="/demo/react/">react (standalone edit)</a><br />&nbsp;</li>
+      <li><a href="/demo/react-preview/">react (preview)</a><br />&nbsp;</li>
       <li>
-        <a href="/demo/react/">React component (live updating)</a><br />&nbsp;
+        <a href="/demo/web-component/">web component (standalone edit)</a
+        ><br />&nbsp;
       </li>
-      <li><a href="/demo/lit/">Web component (old version)</a><br />&nbsp;</li>
+      <li><a href="/demo/lit/">lit (preview)</a><br />&nbsp;</li>
     </ul>
   </body>
 </html>
