+++ packages/editor/src/___dev-demo/main.tsx
@@ -3,8 +3,6 @@ import { createRoot } from 'react-dom/client'
 
 import { Preview } from './preview'
 
-// import './index.css'
-
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
     <Preview />
